Fixed compilation with system Python library.

+++ src/Renderers/Tubes/TriangleTubesCPU.cpp
@@ -79,7 +79,7 @@ void createTriangleTubesRenderDataCPU(
                     vertexDataList);
             lineTangents.push_back(tangent);
             lineNormals.push_back(lastLineNormal);
-            linePointReferenceList.emplace_back(lineId, i);
+            linePointReferenceList.emplace_back(uint32_t(lineId), uint32_t(i));
             numValidLinePoints++;
         }
